test: visual test for button
